+++ mobile/app/src/main/AndroidManifest.xml
@@ -12,20 +12,18 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Team20TimDjordjina">
         <activity
-            android:name=".ResetPasswordActivity"
+            android:name=".activities.HomeActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".activities.ResetPasswordActivity"
             android:exported="false" />
-<!--        <activity-->
-<!--            android:name=".MainActivity"-->
-<!--            android:exported="true">-->
-<!--            <intent-filter>-->
-<!--                <action android:name="android.intent.action.MAIN" />-->
-
-<!--                <category android:name="android.intent.category.LAUNCHER" />-->
-<!--            </intent-filter>-->
-<!--        </activity>-->
+        <activity
+            android:name=".activities.LoginActivity"
+            android:exported="false" />
+         <activity
+             android:name=".activities.HomeActivity"
+             android:exported="false">
+         </activity>
         <activity
             android:name=".activities.RegisterActivity"
             android:exported="false" />
@@ -34,7 +32,6 @@
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
