+++ mobile/app/src/main/java/com/example/team20_tim_djordjina/activities/LoginActivity.java
@@ -3,6 +3,8 @@ package com.example.team20_tim_djordjina.activities;
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Button;
+import android.widget.EditText;
+import android.widget.TextView;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
@@ -14,28 +16,42 @@ import com.example.team20_tim_djordjina.R;
 
 public class LoginActivity extends AppCompatActivity {
 
+    private EditText etEmail, etPassword;
+    private Button btnLogin;
+    private TextView tvForgotPassword, tvError;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
         setContentView(R.layout.activity_login);
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.login_root), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
             return insets;
         });
 
-        Button btnLogin = findViewById(R.id.btnMainLogin);
-        btnLogin.setOnClickListener(v -> {
-            Intent intent = new Intent(MainActivity.this, LoginActivity.class);
-            startActivity(intent);
+        etEmail = findViewById(R.id.etEmail);
+        etPassword = findViewById(R.id.etPassword);
+        tvError = findViewById(R.id.tvError);
+        btnLogin = findViewById(R.id.btnLogin);
+        tvForgotPassword = findViewById(R.id.tvForgotPassword);
+
+        btnLogin.setOnClickListener(view -> {
+            String email = etEmail.getText().toString().trim();
+            String password = etPassword.getText().toString().trim();
+
+            if (email.isEmpty() || password.isEmpty()) {
+                tvError.setText(R.string.email_and_password_cannot_be_empty);
+            }
+//            else {
+//                // log user in
+//            }
         });
 
-// odkomentarisati kada se spoji sa RegisterActivity
-//        Button btnSignup = findViewById(R.id.btnMainSignup);
-//        btnSignup.setOnClickListener(v -> {
-//            Intent intent = new Intent(MainActivity.this, RegisterActivity.class);
-//            startActivity(intent);
-//        });
+        tvForgotPassword.setOnClickListener(v -> {
+            Intent intent = new Intent(LoginActivity.this, ResetPasswordActivity.class);
+            startActivity(intent);
+        });
     }
 }
