+++ mobile/app/src/main/java/com/example/team20_tim_djordjina/activities/ResetPasswordActivity.java
@@ -1,4 +1,4 @@
-package com.example.team20_tim_djordjina;
+package com.example.team20_tim_djordjina.activities;
 
 import android.os.Bundle;
 import android.widget.Button;
@@ -8,6 +8,8 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
+
+import com.example.team20_tim_djordjina.R;
 
 public class ResetPasswordActivity extends AppCompatActivity {
