+++ mobile/app/src/main/java/com/example/team20_tim_djordjina/activities/SplashScreenActivity.java
@@ -2,15 +2,8 @@ package com.example.team20_tim_djordjina.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
-
-import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
-
 import com.example.team20_tim_djordjina.R;
-
 import java.util.Timer;
 import java.util.TimerTask;
 
@@ -25,11 +18,9 @@ public class SplashScreenActivity extends AppCompatActivity {
         new Timer().schedule(new TimerTask() {
             @Override
             public void run() {
-
-                Intent intent = new Intent(SplashScreenActivity.this, RegisterActivity.class);
-                startActivity(intent);
-                finish();
-
+            Intent intent = new Intent(SplashScreenActivity.this, HomeActivity.class);
+            startActivity(intent);
+            finish();
             }
         }, SPLASH_TIME_OUT);
     }
